Refactor Decimal Average Implementation
