+++ velox/functions/prestosql/aggregates/AverageAggregate.cpp
@@ -16,7 +16,8 @@
 #include "velox/exec/Aggregate.h"
 #include "velox/expression/FunctionSignature.h"
 #include "velox/functions/prestosql/aggregates/AggregateNames.h"
-#include "velox/functions/prestosql/aggregates/AverageDecimalAccumulator.h"
+#include "velox/functions/prestosql/aggregates/DecimalAggregate.h"
+#include "velox/type/DecimalUtil.h"
 #include "velox/vector/ComplexVector.h"
 #include "velox/vector/DecodedVector.h"
 #include "velox/vector/FlatVector.h"
@@ -287,268 +288,6 @@ class AverageAggregate : public exec::Aggregate {
   DecodedVector decodedPartial_;
 };
 
-template <typename TUnscaledType>
-class DecimalAverageAggregate : public exec::Aggregate {
- public:
-  explicit DecimalAverageAggregate(TypePtr resultType)
-      : exec::Aggregate(resultType) {}
-
-  int32_t accumulatorFixedWidthSize() const override {
-    return sizeof(AverageDecimalAccumulator);
-  }
-
-  int32_t accumulatorAlignmentSize() const override {
-    return static_cast<int32_t>(sizeof(int128_t));
-  }
-
-  void initializeNewGroups(
-      char** groups,
-      folly::Range<const vector_size_t*> indices) override {
-    setAllNulls(groups, indices);
-    for (auto i : indices) {
-      new (groups[i] + offset_) AverageDecimalAccumulator();
-    }
-  }
-
-  void addRawInput(
-      char** groups,
-      const SelectivityVector& rows,
-      const std::vector<VectorPtr>& args,
-      bool /*mayPushdown*/) override {
-    decodedRaw_.decode(*args[0], rows);
-    if (decodedRaw_.isConstantMapping()) {
-      if (!decodedRaw_.isNullAt(0)) {
-        auto value = decodedRaw_.valueAt<TUnscaledType>(0);
-        rows.applyToSelected(
-            [&](vector_size_t i) { updateNonNullValue(groups[i], value); });
-      }
-    } else if (decodedRaw_.mayHaveNulls()) {
-      rows.applyToSelected([&](vector_size_t i) {
-        if (decodedRaw_.isNullAt(i)) {
-          return;
-        }
-        updateNonNullValue(groups[i], decodedRaw_.valueAt<TUnscaledType>(i));
-      });
-    } else if (!exec::Aggregate::numNulls_ && decodedRaw_.isIdentityMapping()) {
-      auto data = decodedRaw_.data<TUnscaledType>();
-      rows.applyToSelected([&](vector_size_t i) {
-        updateNonNullValue<false>(groups[i], data[i]);
-      });
-    } else {
-      rows.applyToSelected([&](vector_size_t i) {
-        updateNonNullValue(groups[i], decodedRaw_.valueAt<TUnscaledType>(i));
-      });
-    }
-  }
-
-  void addSingleGroupRawInput(
-      char* group,
-      const SelectivityVector& rows,
-      const std::vector<VectorPtr>& args,
-      bool /*mayPushdown*/) override {
-    decodedRaw_.decode(*args[0], rows);
-    if (decodedRaw_.isConstantMapping()) {
-      if (!decodedRaw_.isNullAt(0)) {
-        const auto numRows = rows.countSelected();
-        int64_t overflow = 0;
-        int128_t totalSum{0};
-        auto value = decodedRaw_.valueAt<TUnscaledType>(0);
-        rows.template applyToSelected(
-            [&](vector_size_t i) { updateNonNullValue(group, value); });
-      }
-    } else if (decodedRaw_.mayHaveNulls()) {
-      rows.applyToSelected([&](vector_size_t i) {
-        if (!decodedRaw_.isNullAt(i)) {
-          updateNonNullValue(group, decodedRaw_.valueAt<TUnscaledType>(i));
-        }
-      });
-    } else if (!exec::Aggregate::numNulls_ && decodedRaw_.isIdentityMapping()) {
-      const TUnscaledType* data = decodedRaw_.data<TUnscaledType>();
-      AverageDecimalAccumulator accumulator;
-      rows.applyToSelected([&](vector_size_t i) {
-        accumulator.overflow += DecimalUtil::addWithOverflow(
-            accumulator.sum, data[i].unscaledValue(), accumulator.sum);
-      });
-      accumulator.count = rows.countSelected();
-      char rawData[AverageDecimalAccumulator::serializedSize()];
-      StringView serialized(
-          rawData, AverageDecimalAccumulator::serializedSize());
-      accumulator.serialize(serialized);
-      mergeAccumulators<false>(group, serialized);
-    } else {
-      AverageDecimalAccumulator accumulator;
-      rows.applyToSelected([&](vector_size_t i) {
-        accumulator.overflow += DecimalUtil::addWithOverflow(
-            accumulator.sum,
-            decodedRaw_.valueAt<TUnscaledType>(i).unscaledValue(),
-            accumulator.sum);
-      });
-      accumulator.count = rows.countSelected();
-      char rawData[AverageDecimalAccumulator::serializedSize()];
-      StringView serialized(
-          rawData, AverageDecimalAccumulator::serializedSize());
-      accumulator.serialize(serialized);
-      mergeAccumulators(group, serialized);
-    }
-  }
-
-  void addIntermediateResults(
-      char** groups,
-      const SelectivityVector& rows,
-      const std::vector<VectorPtr>& args,
-      bool /* mayPushdown */) override {
-    decodedPartial_.decode(*args[0], rows);
-    auto intermediateFlatVector =
-        dynamic_cast<const FlatVector<StringView>*>(decodedPartial_.base());
-    if (decodedPartial_.isConstantMapping()) {
-      if (!decodedPartial_.isNullAt(0)) {
-        auto decodedIndex = decodedPartial_.index(0);
-        auto serializedAccumulator =
-            intermediateFlatVector->valueAt(decodedIndex);
-        rows.applyToSelected([&](vector_size_t i) {
-          auto accumulator = decimalAccumulator(groups[i]);
-          accumulator->mergeWith(serializedAccumulator);
-        });
-      }
-    } else if (decodedPartial_.mayHaveNulls()) {
-      rows.applyToSelected([&](vector_size_t i) {
-        if (decodedPartial_.isNullAt(i)) {
-          return;
-        }
-        clearNull(groups[i]);
-        auto decodedIndex = decodedPartial_.index(i);
-        auto serializedAccumulator =
-            intermediateFlatVector->valueAt(decodedIndex);
-        auto accumulator = decimalAccumulator(groups[i]);
-        accumulator->mergeWith(serializedAccumulator);
-      });
-    } else {
-      rows.applyToSelected([&](vector_size_t i) {
-        auto decodedIndex = decodedPartial_.index(i);
-        auto serializedAccumulator =
-            intermediateFlatVector->valueAt(decodedIndex);
-        auto accumulator = decimalAccumulator(groups[i]);
-        accumulator->mergeWith(serializedAccumulator);
-      });
-    }
-  }
-
-  void addSingleGroupIntermediateResults(
-      char* group,
-      const SelectivityVector& rows,
-      const std::vector<VectorPtr>& args,
-      bool /* mayPushdown */) override {
-    decodedPartial_.decode(*args[0], rows);
-    auto intermediateFlatVector =
-        dynamic_cast<const FlatVector<StringView>*>(decodedPartial_.base());
-
-    if (decodedPartial_.isConstantMapping()) {
-      if (!decodedPartial_.isNullAt(0)) {
-        auto decodedIndex = decodedPartial_.index(0);
-        auto serializedAccumulator =
-            intermediateFlatVector->valueAt(decodedIndex);
-        rows.applyToSelected([&](vector_size_t i) {
-          mergeAccumulators(group, serializedAccumulator);
-        });
-      }
-    } else if (decodedPartial_.mayHaveNulls()) {
-      rows.applyToSelected([&](vector_size_t i) {
-        if (decodedPartial_.isNullAt(i)) {
-          return;
-        }
-        auto decodedIndex = decodedPartial_.index(i);
-        auto serializedAccumulator =
-            intermediateFlatVector->valueAt(decodedIndex);
-        mergeAccumulators(group, serializedAccumulator);
-      });
-    } else {
-      rows.applyToSelected([&](vector_size_t i) {
-        auto decodedIndex = decodedPartial_.index(i);
-        auto serializedAccumulator =
-            intermediateFlatVector->valueAt(decodedIndex);
-        mergeAccumulators(group, serializedAccumulator);
-      });
-    }
-  }
-
-  void extractAccumulators(char** groups, int32_t numGroups, VectorPtr* result)
-      override {
-    auto stringViewVector = (*result)->as<FlatVector<StringView>>();
-    stringViewVector->resize(numGroups);
-    uint64_t* rawNulls = nullptr;
-    rawNulls = getRawNulls(stringViewVector);
-    for (auto i = 0; i < numGroups; ++i) {
-      auto accumulator = decimalAccumulator(groups[i]);
-      if (isNull(groups[i])) {
-        stringViewVector->setNull(i, true);
-      } else {
-        clearNull(rawNulls, i);
-        auto size = accumulator->serializedSize();
-        Buffer* buffer = stringViewVector->getBufferWithSpace(size);
-        StringView serialized(buffer->as<char>() + buffer->size(), size);
-        buffer->setSize(buffer->size() + size);
-        accumulator->serialize(serialized);
-        stringViewVector->setNoCopy(i, serialized);
-      }
-    }
-  }
-
-  void extractValues(char** groups, int32_t numGroups, VectorPtr* result)
-      override {
-    auto vector = (*result)->as<FlatVector<TUnscaledType>>();
-    VELOX_CHECK(vector);
-    vector->resize(numGroups);
-    uint64_t* rawNulls = getRawNulls(vector);
-
-    TUnscaledType* rawValues = vector->mutableRawValues();
-    for (int32_t i = 0; i < numGroups; ++i) {
-      char* group = groups[i];
-      if (isNull(group)) {
-        vector->setNull(i, true);
-      } else {
-        clearNull(rawNulls, i);
-        auto accumulator = decimalAccumulator(group);
-        // Handles round-up of fraction results.
-        int128_t average{0};
-        DecimalUtil::computeAverage(
-            average,
-            accumulator->sum,
-            accumulator->count,
-            accumulator->overflow);
-        rawValues[i] = TUnscaledType(average);
-      }
-    }
-  }
-
-  template <bool tableHasNulls = true>
-  void mergeAccumulators(char* group, const StringView& serialized) {
-    if constexpr (tableHasNulls) {
-      exec::Aggregate::clearNull(group);
-    }
-    auto accumulator = decimalAccumulator(group);
-    accumulator->mergeWith(serialized);
-  }
-
-  template <bool tableHasNulls = true>
-  void updateNonNullValue(char* group, TUnscaledType value) {
-    if constexpr (tableHasNulls) {
-      exec::Aggregate::clearNull(group);
-    }
-    auto accumulator = decimalAccumulator(group);
-    accumulator->overflow += DecimalUtil::addWithOverflow(
-        accumulator->sum, value.unscaledValue(), accumulator->sum);
-    accumulator->count += 1;
-  }
-
- private:
-  inline AverageDecimalAccumulator* decimalAccumulator(char* group) {
-    return exec::Aggregate::value<AverageDecimalAccumulator>(group);
-  }
-
-  DecodedVector decodedRaw_;
-  DecodedVector decodedPartial_;
-};
-
 void checkSumCountRowType(TypePtr type, const std::string& errorMessage) {
   VELOX_CHECK(
       type->kind() == TypeKind::ROW || type->kind() == TypeKind::VARBINARY,
@@ -565,6 +304,22 @@ void checkSumCountRowType(TypePtr type, const std::string& errorMessage) {
   VELOX_CHECK_EQ(
       type->childAt(1)->kind(), TypeKind::BIGINT, "{}", errorMessage);
 }
+
+template <typename TUnscaledType>
+class DecimalAverageAggregate : public DecimalAggregate<TUnscaledType> {
+ public:
+  explicit DecimalAverageAggregate(TypePtr resultType)
+      : DecimalAggregate<TUnscaledType>(resultType) {}
+
+  virtual TUnscaledType computeFinalValue(
+      LongDecimalWithOverflowState* accumulator) final {
+    // Handles round-up of fraction results.
+    int128_t average{0};
+    DecimalUtil::computeAverage(
+        average, accumulator->sum, accumulator->count, accumulator->overflow);
+    return TUnscaledType(average);
+  }
+};
 
 bool registerAverageAggregate(const std::string& name) {
   std::vector<std::shared_ptr<exec::AggregateFunctionSignature>> signatures;
@@ -652,7 +407,12 @@ bool registerAverageAggregate(const std::string& name) {
               if (inputType->kind() == TypeKind::LONG_DECIMAL) {
                 return std::make_unique<
                     DecimalAverageAggregate<UnscaledLongDecimal>>(resultType);
-              } else {
+              } else if (
+                  inputType->kind() == TypeKind::SHORT_DECIMAL ||
+                  inputType->kind() == TypeKind::VARBINARY) {
+                // If the input and out type are VARBINARY, then the
+                // LongDecimalWithOverflowState is used and the template type
+                // does not matter.
                 return std::make_unique<
                     DecimalAverageAggregate<UnscaledShortDecimal>>(resultType);
               }
